QSPP::startAsyncRead(): set a correct type for variable
'bytesToRead' takes the result of signed 64-bit computation which could
be negative. So, its type should be qint64.

+++ src/serialport/qserialport_win.cpp
@@ -541,7 +541,7 @@ bool QSerialPortPrivate::startAsyncRead()
     if (readStarted)
         return true;
 
-    DWORD bytesToRead = ReadChunkSize;
+    qint64 bytesToRead = ReadChunkSize;
 
     if (readBufferMaxSize && bytesToRead > (readBufferMaxSize - buffer.size())) {
         bytesToRead = readBufferMaxSize - buffer.size();
